+++ alicompat/forge/src/compat/farlanders/java/com/yanny/alicompat/compat/farlanders/FarlandersCompat.java
@@ -6,8 +6,14 @@ import com.yanny.ali.api.IServerRegistry;
 import com.yanny.ali.api.TradeLevelInfo;
 import com.yanny.alicompat.IModCompat;
 import com.yanny.alicompat.accessor.PluginUtils;
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.npc.VillagerTrades;
 import org.jetbrains.annotations.NotNull;
+
+import java.util.function.IntFunction;
+import java.util.function.Supplier;
 
 public class FarlandersCompat implements IModCompat {
     private static final String MOD_ID = "farlanders";
@@ -22,8 +28,14 @@ public class FarlandersCompat implements IModCompat {
     public void registerServer(IServerRegistry registry) {
         PluginUtils.registerItemListing(registry, FarlanderTrades.Trade.class, TradeAccessor.class);
 
-        registry.registerTrades(new ResourceLocation(MOD_ID, "farlander"), () -> FarlanderTrades.FARLANDER_TRADES, (level) -> new TradeLevelInfo(new RangeValue(2)));
-        registry.registerTrades(new ResourceLocation(MOD_ID, "elder_farlander"), () -> FarlanderTrades.ELDER_TRADES, (level) -> new TradeLevelInfo(new RangeValue(2)));
-        registry.registerTrades(new ResourceLocation(MOD_ID, "wanderer"), () -> FarlanderTrades.WANDERER_TRADES, (level) -> new TradeLevelInfo(new RangeValue(5)));
+        registerTrades(registry, "farlander", () -> FarlanderTrades.FARLANDER_TRADES, (level) -> new TradeLevelInfo(new RangeValue(2)));
+        registerTrades(registry, "elder_farlander", () -> FarlanderTrades.ELDER_TRADES, (level) -> new TradeLevelInfo(new RangeValue(2)));
+        registerTrades(registry, "wanderer", () -> FarlanderTrades.WANDERER_TRADES, (level) -> new TradeLevelInfo(new RangeValue(5)));
+    }
+
+    private static void registerTrades(IServerRegistry registry, String name, Supplier<Int2ObjectMap<VillagerTrades.ItemListing[]>> itemListings, IntFunction<TradeLevelInfo> levelInfo) {
+        ResourceLocation traderId = new ResourceLocation(MOD_ID, name);
+
+        registry.registerTrades(traderId, BuiltInRegistries.ENTITY_TYPE.get(traderId), itemListings, levelInfo);
     }
 }