+++ ali/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Added support for custom ingredients that are not Ingredient subclasses
 - Improved count and chance detection
 - Fixed server crash on Forge when GLM cannot provide its codec
+- Reworked trades UI, displaying trader spawn egg
+- Fixed missing tooltip in JEI and REI block loot
 
 ## [2.2.0]